Symlinks improvement

+++ src/actions/RemoveACL.php
@@ -13,6 +13,8 @@ use eseperio\filescatalog\models\AccessControl;
 use eseperio\filescatalog\models\InodePermissionsForm;
 use Yii;
 use yii\base\Action;
+use yii\web\ForbiddenHttpException;
+use yii\web\ServerErrorHttpException;
 
 class RemoveACL extends Action
 {
@@ -20,18 +22,23 @@ class RemoveACL extends Action
     {
         $permModel = new InodePermissionsForm();
         $permModel->scenario = AccessControl::SCENARIO_DELETE;
-        $permModel->load(Yii::$app->request->post(), '');
-        if ($permModel->validate())
+        $permModel->setAttributes(Yii::$app->request->post(), false);
+        if ($permModel->validate()) {
             try {
+
                 $realModel = AccessControl::find()->where([
                     'user_id' => $permModel->user_id,
                     'role' => $permModel->role,
                     'inode_id' => $permModel->inode_id
                 ])->one();
-                $realModel->delete();
+                if (!empty($realModel))
+                    $realModel->delete();
             } catch (\Throwable $e) {
                 throw $e;
             }
+        } else {
+            throw new ServerErrorHttpException(Yii::t('filescatalog', 'An error ocurred deleting this item'));
+        }
 
 
         return $this->controller->goBack();
